image: allow putting `/boot` on the rootfs
For most devices, we want a separate `/boot` so, for example, the
bootloader can fetch the initramfs from an unencrypted partition.
However, there might be cases where this isn't necessary, or it can make
things overly complicated.

This change adds a new template variable for `partition-*` sub-recipes,
allowing the whole system to use a single partition, as well as the
corresponding changes to the `image` recipe and `setup-luks` script.