qcom: use a single rootfs partition including `/boot`
We used to keep `/boot` separated so we could at some point enable FDE
with `osk-sdl` (which results in a very large initramfs, too large to be
included in the boot image). However, this requires finding a usable
partition to flash `/boot` onto, which is not always feasible. Moreover,
we recently switched to using `unl0kr` which doesn't increase the size
of the initramfs too much.

Therefore, we can get rid of the "2 partitions images" for this device
family and simply put `/boot` back on the rootfs.