image: fix encrypted builds
`debos`, when running under `fakemachine`, doesn't expose the whole
`/dev` folder, but only the block devices it knows about (`/dev/disk`
and `/dev/vd*`). When using an encrypted rootfs, this causes issues as
`/dev/mapper/root` and the corresponding `/dev/dm-*` nodes aren't known
to `debos`, and therefore not available within the chroot. As a
consequence, `cryptsetup` is unable to determine the proper parameters
and leaves an empty `crypttab` in the initramfs, leading to unusable
images.

A manual update of the initramfs is therefore needed, ensuring the whole
`/dev` is accessible from the chroot so `crypttab` can be generated
properly.

This part was accidentally removed in 8d5deb72 so let's restore it with
an additional comment explaining its importance.