Make `qcom` devices more configurable
Currently a lot of information is hardcoded in the various scripts,
including the SoC identifier, partition sector size, addresses for the
boot images and vendor/model/variant information.

This change gets all this information out of the scripts/recipes and
into a per-SoC config file, so it gets easier to support new devices in
the future.

Fixes: #16