Add `qcom-wip` device type to help supporting new devices
!31 was an attempt at creating a dummy device type in order to make it
easier to support new devices. However, by trying to be completely
generic, it was overly complex and consequently was never actually
finished.

In the real world, devices likely to be supported are only qcom-based
ones (at least for now), so we shouldn't bother with addressing
non-existent use-cases and instead provide a simple-ish way of building
images for new qcom devices.

This is achieved by providing a new `qcom-wip` device type, which does
the following:
* use the generic `qcom-support-common` instead of `<soc>-support`
* install all `.deb` packages found under `devices/qcom/packages`
* copy any droid-juicer config file under `devices/qcom/droid-juicer` to
  the image

With those changes, building an image for a yet-unsupported device
requires the following steps:
* build a `.deb` kernel package for this device (using `make bindeb-pkg`
  on a downstream tree, for example) and copy it to the
  `devices/qcom/packages` folder
* fill in the `devices/qcom/configs/wip.toml` file with the appropriate
  information for the target device
* optional: create a `droid-juicer` config named `<compatible>.toml` and
  copy it to the `devices/qcom/droid-juicer` folder
* run `./build.sh -t qcom-wip`

Those images also include `openssh-server` by default in order to ease
debugging.