devices: qcom: bootloader.sh: switch to using `mkbootimg`
`abootimg` is easier to use, but it's an old piece of software and has
been unmaintained for many years. As such, it can't handle modern
variants and features of the bootimg format which we need for supporting
new devices in the near future.

Let's switch to using `mkbootimg` instead, which is an official Android
tool and, as such, still actively maintained and up-to-date.