scripts: add unstable/exp to work around packages from testing
Some critical packages (`iio-sensor-proxy`, `squeekboard` for Phosh
images...) are not longer in testing due to unresolved Debian bugs. As
those are still present in `unstable` (or `experimental` for some),
let's ensure the image build process has access to those repos.

Priorities are set so that the package from `trixie` is installed if it
exists. Otherwise the version of `experimental` is used, unless the
package only exists in `unstable`.

The corresponding sources/preferences files are then removed during
image cleanup.