devices: qcom: bootloader: set loglevel if a console is defined
Some devices have a `console=...` kernel param in order to work around a
race condition in the display's initialization. However, this problem's
discussion mentions[1] that `loglevel` should be high enough for this
workaround to be effective. We therefore set it to 7 (debug) whenever
the kernel cmdline contains a `console=...` argument.

[1] https://gitlab.freedesktop.org/drm/msm/-/issues/46#note_2213313