devices: amd64: add script to help sync firmware list with Debian
For x86 devices, we want to support the same devices as Debian do; we
already use the Debian kernel here, but are missing some of the firmware
packages shipped by default.

This change adds a small script that fetches the full list of packages
from the latest testing installer build and extracts the name of all
firmware packages, formatted so they can easily be copy/pasted into the
recipe's package list.

Take this opportunity to update this list for current images.